Trait tower_web::error::IntoCatch

pub trait IntoCatch<S> {
    type Catch: Catch;
    fn into_catch(self) -> Self::Catch;
}

Convert Self into a value that implements Catch.

By implementing IntoCatch for a type, you define how it will be converted into a Catch instance.

The Serializer configured for the service is provided. This allows the Catch instance to be able to serialize responses.

Associated Types

type Catch: Catch

The Catch value being converted to

Loading content...
Required methods

fn into_catch(self) -> Self::Catch

Convert self into a Catch value.

Loading content...