Module tower_web::extract

Extract data from the HTTP request.

The Extract trait is implemented by types that can be constructed by using data from an HTTP request. Resource method argument types must implement Extract. This is how impl_web! is able to instantiate them using the request. See library level documentation for more details.

Currently, Extract implementations are provided for the following types:

Bytes

Option

PathBuf

String

Vec<u8>

More implementations can be added by submitting a PR.

Also, Extract can be implemented for custom types by using the [derive(Extract)] proc macro. See library level documentation for more details.

Modules

http_date_time
HTTP combined date and time value.

option
Types used to extract Option values from an HTTP request.

Structs

Context
Context available when extracting data from the HTTP request.

Error
Errors that can happen while extracting data from an HTTP request.

Immediate
Implements ExtractFuture such that the result is immediately available.

Traits

Extract
Extract a value from an HTTP request.

ExtractFuture
Future representing the completion of extracting a value from a request